put rewrite rules behind inductive type, so they won't get fired otherwise
Note that this screws up a bunch of theorems that were working via crush, as
it (doesnt seem to) then be able to use the extern hints. There are now a couple
of manual `rewrite (unRR Hintn)` tactics, which obviously should be removed,
probably by eliminating crush and instead using / writing tactics that
understand this local hint style better.